Assert authenticated credential is stored
